requirements.txt
Jinja2~=2.11
PyYAML~=5.3
attrs~=19.3
dill~=0.3
pandas~=1.0
parse~=1.16
pathos~=0.2
pdoc3~=0.8
plumbum~=1.6
psutil~=5.7
psycopg2-binary~=2.8
pygit2>=1.2
pygtrie~=2.3
pyparsing~=2.4
sqlalchemy-migrate~=0.13
virtualenv~=20.0